Snake and Ladder Online Gaming
This a game played by two to four players. It is embedded online as it is built using applet.The players roll the dice and based on the number they get they move up the board.The Player who reaches the finish point first wins. This Game has been built using JSP.It also enables full screen support.